% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getAbsoluteAbundance}
\alias{getAbsoluteAbundance}
\alias{getAbsoluteAbundance,CrcBiomeScreen-method}
\title{Accessor for AbsoluteAbundance slot of CrcBiomeScreen object}
\usage{
getAbsoluteAbundance(object)

\S4method{getAbsoluteAbundance}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with A data.frame containing
absolute abundance data.
}
\description{
Accessor for AbsoluteAbundance slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getAbsoluteAbundance(CrcBiomeScreen)}: Retrieve absolute abundance data
from a CrcBiomeScreen object.

}}
\examples{
rel_abund <- data.frame(Sample1 = 10, Sample2 = 20)
rownames(rel_abund) <- "TaxaA"

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA", row.names = "TaxaA"),
  SampleData = data.frame(
    number_reads = c(10000, 10000),
    condition = c("control", "CRC"),
    row.names = c("Sample1", "Sample2")
  )
)

# Retrieve absolute abundance
getAbsoluteAbundance(toy_obj)
}
