% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getModelResult}
\alias{getModelResult}
\alias{getModelResult,CrcBiomeScreen-method}
\title{Accessor for ModelResult slot of CrcBiomeScreen object}
\usage{
getModelResult(object)

\S4method{getModelResult}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a list containing fitted model results.
}
\description{
Accessor for ModelResult slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getModelResult(CrcBiomeScreen)}: Retrieve model results
from a CrcBiomeScreen object.

}}
\examples{
rel_abund <- data.frame(Sample1 = 10, Sample2 = 20)
rownames(rel_abund) <- "TaxaA"

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA", row.names = "TaxaA"),
  SampleData = data.frame(
    number_reads = c(10000, 10000),
    condition = c("control", "CRC"),
    row.names = c("Sample1", "Sample2")
  )
)

getModelData(toy_obj)
}
