% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getOutlierSamples}
\alias{getOutlierSamples}
\alias{getOutlierSamples,CrcBiomeScreen-method}
\title{Accessor for OutlierSamples}
\usage{
getOutlierSamples(object)

\S4method{getOutlierSamples}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A A \code{CrcBiomeScreen} object. object.}
}
\value{
A character vector of detected outlier sample IDs, or NULL if no outliers have been recorded.
}
\description{
Accessor for OutlierSamples
}
\section{Methods (by class)}{
\itemize{
\item \code{getOutlierSamples(CrcBiomeScreen)}: results
from a CrcBiomeScreen object.

}}
\examples{
# getOutlierSamples() is typically used after running qcByCmdscale().
# See the package vignette for a complete workflow example.
}
