% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getRelativeAbundance}
\alias{getRelativeAbundance}
\alias{getRelativeAbundance,CrcBiomeScreen-method}
\title{Accessor for RelativeAbundance slot of CrcBiomeScreen object}
\usage{
getRelativeAbundance(object)

\S4method{getRelativeAbundance}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a data.frame containing relative abundance data.
}
\description{
Accessor for RelativeAbundance slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getRelativeAbundance(CrcBiomeScreen)}: Retrieve relative abundance data
from a CrcBiomeScreen object.

}}
\examples{
rel_abund <- data.frame(Sample1 = 10, Sample2 = 20)
rownames(rel_abund) <- "TaxaA"

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA", row.names = "TaxaA"),
  SampleData = data.frame(
    number_reads = c(10000, 10000),
    condition = c("control", "CRC"),
    row.names = c("Sample1", "Sample2")
  )
)

getRelativeAbundance(toy_obj)
}
