% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getSampleData}
\alias{getSampleData}
\alias{getSampleData,CrcBiomeScreen-method}
\alias{getModelData,CrcBiomeScreen-method}
\title{Accessor for SampleData slot of CrcBiomeScreen object}
\usage{
getSampleData(object)

\S4method{getSampleData}{CrcBiomeScreen}(object)

\S4method{getModelData}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a data.frame containing sample metadata.
}
\description{
Accessor for SampleData slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getSampleData(CrcBiomeScreen)}: Retrieve sample metadata
from a CrcBiomeScreen object.

\item \code{getModelData(CrcBiomeScreen)}: Retrieve sample metadata
from a CrcBiomeScreen object.

}}
\examples{
rel_abund <- data.frame(Sample1 = 10, Sample2 = 20)
rownames(rel_abund) <- "TaxaA"

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA", row.names = "TaxaA"),
  SampleData = data.frame(
    number_reads = c(10000, 10000),
    condition = c("control", "CRC"),
    row.names = c("Sample1", "Sample2")
  )
)

getSampleData(toy_obj)
}
