% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeepGenusLevel.R
\name{KeepGenusLevel}
\alias{KeepGenusLevel}
\title{Keep the taxa data at genus level}
\usage{
KeepGenusLevel(CrcBiomeScreenObject)
}
\arguments{
\item{CrcBiomeScreenObject}{}
}
\value{
CrcBiomeScreenObject$GenusLevelData
}
\description{
Keep the taxa data at genus level
}
\examples{
CrcBiomeScreenObject <- KeepGenusLevel(CrcBiomeScreenObject)
}
