% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_functions}
\alias{load_functions}
\title{Load all functions for testing}
\usage{
load_functions()
}
\description{
This function sources all R files in the current directory
for testing purposes. This is not meant for the final package.
}
