% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingRF.R
\name{ModelingRF}
\alias{ModelingRF}
\title{The packaging function for Random Forest modeling}
\usage{
ModelingRF(
  CrcBiomeScreenObject = NULL,
  k.rf = n_cv,
  TaskName = NULL,
  TrueLabel = NULL,
  num_cores = NULL
)
}
\arguments{
\item{CrcBiomeScreenObject}{A \code{CrcBiomeScreenObject} containing normalized microbiome data, sample metadata, etc.}

\item{k.rf}{Set the number of cross validation}

\item{TaskName}{A character string used to label the output}

\item{TrueLabel}{This label is the future prediction target}

\item{num_cores}{Set the number of the cores in parallel computing}
}
\value{
A A \code{CrcBiomeScreen} object. with the modelling results.
}
\description{
The packaging function for Random Forest modeling
}
\examples{
# Minimal runnable example illustrating required inputs for ModelingRF

# Create toy relative abundance matrix
rel_abund <- data.frame(S1 = 10, S2 = 20)
rownames(rel_abund) <- "TaxaA"

# Create sample metadata
sample_info <- data.frame(
  number_reads = c(10000, 12000),
  condition = c("control", "CRC"),
  row.names = c("S1", "S2")
)

# Construct minimal CrcBiomeScreen object
obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = sample_info,
)


# NOT RUN: Actual model fitting is time-consuming
# out <- ModelingRF(
#   CrcBiomeScreenObject = obj,
#   k.rf = 2,
#   TaskName = "toy_RF",
#   TrueLabel = c("control", "CRC"),
#   num_cores = 1
# )

# The example instead demonstrates setup only
obj

}
