% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingXGBoost.R
\name{ModelingXGBoost}
\alias{ModelingXGBoost}
\title{The packaging function for XGBoost modeling}
\usage{
ModelingXGBoost(
  CrcBiomeScreenObject = NULL,
  k.rf = 10,
  repeats = 5,
  TaskName = NULL,
  TrueLabel = NULL,
  num_cores = num_cores
)
}
\arguments{
\item{CrcBiomeScreenObject}{A \code{CrcBiomeScreenObject} containing normalized microbiome data, sample metadata, etc.}

\item{k.rf}{Set the number of cross validation}

\item{repeats}{Set the number of repeats in cross validation}

\item{TaskName}{A character string used to label the output}

\item{TrueLabel}{This label is the future prediction target}

\item{num_cores}{Set the number of the cores in parallel computing}
}
\value{
CrcBiomeScreenObject
}
\description{
The packaging function for XGBoost modeling
}
\examples{
# Minimal runnable example for ModelingXGBoost

rel_abund <- data.frame(S1 = 10, S2 = 20)
rownames(rel_abund) <- "TaxaA"

sample_info <- data.frame(
  number_reads = c(10000, 12000),
  condition = c("control", "CRC"),
  row.names = c("S1", "S2")
)

obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = sample_info
)

obj@ModelData <- list(
  Training = data.frame(x = c(1, 2)),
  TrainLabel = factor(c("control", "CRC"))
)

# out <- ModelingXGBoost(
# CrcBiomeScreenObject = obj,
#  k.rf = 2,
# TaskName = "toy_XGB",
# TrueLabel = c("control", "CRC"),
#  num_cores = 1
# )

out
}
