% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingXGBoost_noweights.R
\name{ModelingXGBoost_noweights}
\alias{ModelingXGBoost_noweights}
\title{The packaging function for XGBoost modeling without using class weights}
\usage{
ModelingXGBoost_noweights(
  CrcBiomeScreenObject = NULL,
  k.rf = 10,
  repeats = 5,
  TaskName = NULL,
  TrueLabel = NULL,
  num_cores = num_cores
)
}
\arguments{
\item{CrcBiomeScreenObject}{A \code{CrcBiomeScreenObject} containing normalized microbiome data, sample metadata, etc.}

\item{k.rf}{Set the number of cross validation}

\item{repeats}{Set the number of repeats in cross validation}

\item{TaskName}{A character string used to label the output}

\item{TrueLabel}{This label is the future prediction target}

\item{num_cores}{Set the number of the cores in parallel computing}
}
\value{
A \linkS4class{CrcBiomeScreenObject} with the modelling results.
}
\description{
The packaging function for XGBoost modeling without using class weights
}
\examples{
# Minimal runnable example for ModelingXGBoost_noweights

rel_abund <- data.frame(S1 = 10, S2 = 20)
rownames(rel_abund) <- "TaxaA"

sample_info <- data.frame(
  number_reads = c(10000, 12000),
  condition = c("control", "CRC"),
  row.names = c("S1", "S2")
)

obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = sample_info
)


# out <- ModelingXGBoost_noweights(
# CrcBiomeScreenObject = obj,
# k.rf = 2,
# TaskName = "toy_XGB_nw",
# TrueLabel = c("control", "CRC"),
# num_cores = 1
# )

obj
}
