% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{setNormalizedData-setter}
\alias{setNormalizedData-setter}
\alias{setNormalizedData<-}
\alias{setNormalizedData<-,CrcBiomeScreen-method}
\title{setNormalizedData<-: Setter for NormalizedData slot
of CrcBiomeScreen object}
\usage{
setNormalizedData(object) <- value

\S4method{setNormalizedData}{CrcBiomeScreen}(object) <- value
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}

\item{value}{A data.frame or matrix containing normalized abundance data.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a  modified \linkS4class{CrcBiomeScreen} object.
}
\description{
setNormalizedData<-: Setter for NormalizedData slot
of CrcBiomeScreen object
}
\section{Functions}{
\itemize{
\item \code{setNormalizedData(CrcBiomeScreen) <- value}: Replace the NormalizedData slot
of a CrcBiomeScreen object.

}}
\examples{
rel_abund <- data.frame(Sample1 = 10, Sample2 = 20)
rownames(rel_abund) <- "TaxaA"

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA", row.names = "TaxaA"),
  SampleData = data.frame(
    number_reads = c(10000, 10000),
    condition = c("control", "CRC"),
    row.names = c("Sample1", "Sample2")
  )
)

setNormalizedData(toy_obj) <- data.frame(n1 = 1:2)
getNormalizedData(toy_obj)
}
