% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{setTaxaData-setter}
\alias{setTaxaData-setter}
\alias{setTaxaData<-}
\alias{setTaxaData<-,CrcBiomeScreen-method}
\title{setTaxaData<-: Setter for TaxaData slot of CrcBiomeScreen object}
\usage{
setTaxaData(object) <- value

\S4method{setTaxaData}{CrcBiomeScreen}(object) <- value
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}

\item{value}{A data.frame containing updated taxonomic annotations.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a modified \linkS4class{CrcBiomeScreen} object.
}
\description{
setTaxaData<-: Setter for TaxaData slot of CrcBiomeScreen object
}
\section{Functions}{
\itemize{
\item \code{setTaxaData(CrcBiomeScreen) <- value}: Replace the TaxaData slot
of a CrcBiomeScreen object.

}}
\examples{
rel_abund <- data.frame(Sample1 = 10, Sample2 = 20)
rownames(rel_abund) <- "TaxaA"

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA", row.names = "TaxaA"),
  SampleData = data.frame(
    number_reads = c(10000, 10000),
    condition = c("control", "CRC"),
    row.names = c("Sample1", "Sample2")
  )
)

setTaxaData(toy_obj) <- data.frame(Taxa = "NewTaxa")
getTaxaData(toy_obj)
}
