% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateModel.R
\name{TaskName Save the running results}
\alias{TaskName Save the running results}
\alias{EvaluateModel}
\title{Test the Model}
\usage{
EvaluateModel(
  CrcBiomeScreenObject = NULL,
  model_type = c("RF", "XGBoost"),
  TaskName = NULL,
  TrueLabel = NULL,
  PlotAUC = NULL
)
}
\arguments{
\item{ModelData}{List for the Datasets}

\item{label}{The label are used for modeling}

\item{taxa_col}{Select the needed Column}

\item{best.params}{The best parameters for model}
}
\description{
Test the Model
}
