% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZellerG_2014_RelativeAbundance.R
\docType{data}
\name{ZellerG_2014_RelativeAbundance}
\alias{ZellerG_2014_RelativeAbundance}
\title{ZellerG_2014_RelativeAbundance}
\format{
A data frame with ...
}
\source{
curatedMetagenomicData package, ZellerG et al. (2014)
}
\usage{
ZellerG_2014_RelativeAbundance
}
\value{
A data frame (or matrix) of microbial relative abundance values from
Zeller et al. (2014), with taxa as rows and samples as columns.
}
\description{
A small subset of ZellerG_2014_RelativeAbundance microbiome data.
}
\keyword{datasets}
