% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClassBalance.R
\name{checkClassBalance}
\alias{checkClassBalance}
\title{Check the sample distribution of the dataset and
give the suggestion if need the class weight or not}
\usage{
checkClassBalance(labels, outdir = tempdir(), threshold = 0.5, plot = TRUE)
}
\arguments{
\item{labels}{The label for distribution}

\item{outdir}{The output directory where plots will be saved
(default: tempdir()).}

\item{threshold}{The threshold for the ratio (0.5) if it is the
imbalanced dataset}

\item{plot}{Choose to have the figures or not}
}
\value{
A A \code{CrcBiomeScreen} object. object with updated slots.
}
\description{
Check the sample distribution of the dataset and
give the suggestion if need the class weight or not
}
\examples{
# Small toy example for runnable demonstration
train_labels <- factor(c("control", "CRC", "control", "CRC"))
checkClassBalance(train_labels)
}
