% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClassBalance.R
\name{checkClassBalance}
\alias{checkClassBalance}
\title{Check the sample distribution of the dataset and
give the suggestion if need the class weight or not}
\usage{
checkClassBalance(labels, threshold = 0.5, plot = TRUE)
}
\arguments{
\item{labels}{The label for distribution}

\item{threshold}{The threshold for the ratio (0.5) if it is the imbalanced dataset}

\item{plot}{Choose to have the figures or not}
}
\value{
A list including the details about this step
}
\description{
Check the sample distribution of the dataset and
give the suggestion if need the class weight or not
}
\examples{
checkClassBalance(CrcBiomeScreenObject$ModelData$TrainLabel)

}
