% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{CreateCrcBiomeScreenObject}
\alias{CreateCrcBiomeScreenObject}
\title{Create the CrcBiomeScreenObject for analysis}
\usage{
CreateCrcBiomeScreenObject(
  AbsoluteAbundance = NULL,
  TaxaData = NULL,
  SampleData = NULL,
  RelativeAbundance = NULL
)
}
\arguments{
\item{AbsoluteAbundance}{A numeric matrix or data frame containing absolute abundance data}

\item{TaxaData}{A data frame containing taxonomic information for each feature}

\item{SampleData}{A data frame containing sample metadata}

\item{RelativeAbundance}{A numeric matrix or data frame containing relative abundance data}
}
\value{
An object with data
}
\description{
Creates the CrcBiomeScreenObject to contain the data for analysis.
}
\examples{
if (requireNamespace("curatedMetagenomicData", quietly = TRUE)) {
  toydata <- curatedMetagenomicData::curatedMetagenomicData(
    "ThomasAM_2018a.relative_abundance",
    dryrun = FALSE, rownames = "short"
  )}
  CrcBiomeScreenObject <- CreateCrcBiomeScreenObject(
    RelativeAbundance = toydata[[1]]@assays@data@listData$relative_abundance,
    TaxaData = toydata[[1]]@rowLinks$nodeLab,
    SampleData = toydata[[1]]@colData
  )
}
