% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromTSE.R
\name{CreateCrcBiomeScreenObjectFromTSE}
\alias{CreateCrcBiomeScreenObjectFromTSE}
\title{Create a CrcBiomeScreen object from TreeSummarizedExperiment}
\usage{
CreateCrcBiomeScreenObjectFromTSE(tse, assay_name = NULL)
}
\arguments{
\item{tse}{A TreeSummarizedExperiment object containing microbiome data.}

\item{assay_name}{Which assay to use (default: "counts" or "relative_abundance").}
}
\value{
A \linkS4class{CrcBiomeScreen} object.
}
\description{
Wrapper function to directly convert a TreeSummarizedExperiment object
into a \linkS4class{CrcBiomeScreen} S4 object for downstream analysis.
}
\examples{
if (requireNamespace("curatedMetagenomicData", quietly = TRUE)) {
  tse <- curatedMetagenomicData::curatedMetagenomicData(
    "ThomasAM_2018a.relative_abundance",
    dryrun = FALSE, rownames = "short"
  )[[1]]
  obj <- CreateCrcBiomeScreenObjectFromTSE(tse)
  obj
}
}
