% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateModel.R
\name{EvaluateModel}
\alias{EvaluateModel}
\title{Evaluate the model to select the optimal model}
\usage{
EvaluateModel(
  CrcBiomeScreenObject = NULL,
  model_type = c("RF", "XGBoost"),
  TaskName = NULL,
  TrueLabel = NULL,
  PlotAUC = NULL
)
}
\arguments{
\item{CrcBiomeScreenObject}{A CrcBiomeScreenObject containing the model data and results}

\item{model_type}{A character vector indicating the type of model to evaluate. Options are "RF" for Random Forest and "XGBoost" for XGBoost.}

\item{TaskName}{A character string used to label the output files and results.}

\item{TrueLabel}{The true label for the classification task, which is used to evaluate the model's performance.}

\item{PlotAUC}{A logical value indicating whether to plot the AUC curve. If TRUE, the AUC curve will be saved as a PDF file.}
}
\value{
A CrcBiomeScreenObject with the evaluation results stored in the \code{EvaluateResult} slot.
}
\description{
Evaluate the model to select the optimal model
}
\examples{
CrcBiomeScreenObject <- EvaluateModel(CrcBiomeScreenObject,
  model_type = "RF",
  TaskName = "ToyData_RF_Test",
  TrueLabel = "CRC",
  PlotAUC = TRUE
)

}
