% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateModel.R
\name{EvaluateXGBoost}
\alias{EvaluateXGBoost}
\title{Evaluate the XGBoost model}
\usage{
EvaluateXGBoost(
  CrcBiomeScreenObject = NULL,
  TaskName = NULL,
  TrueLabel = NULL,
  PlotAUC = NULL
)
}
\arguments{
\item{CrcBiomeScreenObject}{A CrcBiomeScreenObject containing the model data and results}

\item{TaskName}{A character string used to label the output files and results.}

\item{TrueLabel}{The true label for the classification task, which is used to evaluate the model's performance.}

\item{PlotAUC}{A logical value indicating whether to plot the AUC curve. If TRUE, the AUC curve will be saved as a PDF file.}
}
\value{
A CrcBiomeScreenObject with the evaluation results stored in the \code{EvaluateResult$XGBoost} slot.
}
\description{
Evaluate the XGBoost model
}
\examples{
CrcBiomeScreenObject <- EvaluateXGBoost(
  CrcBiomeScreenObject = CrcBiomeScreenObject,
  TaskName = TaskName,
  TrueLabel = TrueLabel,
  PlotAUC = PlotAUC
)

}
