% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterDataset.R
\name{FilterDataSet}
\alias{FilterDataSet}
\title{Filter the CrcBiomeScreenObject dataset based on a specific label}
\usage{
FilterDataSet(
  CrcBiomeScreenObject = NULL,
  label = NULL,
  condition_col = "study_condition"
)
}
\arguments{
\item{CrcBiomeScreenObject}{A \code{CrcBiomeScreenObject} containing normalized microbiome data and sample metadata.}

\item{label}{A character vector specifying the label(s) to filter the dataset by.}

\item{condition_col}{A character string indicating the column in the SampleData that contains the condition labels (default is "study_condition").}
}
\value{
A \code{CrcBiomeScreenObject} with filtered data based on the specified label.
}
\description{
Filter the CrcBiomeScreenObject dataset based on a specific label
}
\examples{
ValidationData_filtered <- FilterDataSet(ValidationData,
  label = c("CRC", "control"),
  condition_col = "study_condition"
)

}
