% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getAbsoluteAbundance}
\alias{getAbsoluteAbundance}
\alias{getAbsoluteAbundance,CrcBiomeScreen-method}
\title{Accessor for AbsoluteAbundance slot of CrcBiomeScreen object}
\usage{
getAbsoluteAbundance(object)

\S4method{getAbsoluteAbundance}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A A \code{CrcBiomeScreen} object. object.}
}
\value{
A A \code{CrcBiomeScreen} object. object with A data.frame containing
absolute abundance data.
}
\description{
Accessor for AbsoluteAbundance slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getAbsoluteAbundance(CrcBiomeScreen)}: Retrieve absolute abundance data
from a CrcBiomeScreen object.

}}
\examples{
# Construct minimal example object
toy_obj <- CreateCrcBiomeScreenObject(
  AbsoluteAbundance = data.frame(TaxaA = c(1000, 2000)),
  RelativeAbundance = data.frame(TaxaA = c(10, 20)),
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = data.frame(
    number_reads = 10000,
    condition = "control"
  )
)

# Retrieve absolute abundance
getAbsoluteAbundance(toy_obj)
}
