% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getAbsoluteAbundance}
\alias{getAbsoluteAbundance}
\alias{getAbsoluteAbundance,CrcBiomeScreen-method}
\title{Accessor for AbsoluteAbundance slot of CrcBiomeScreen object}
\usage{
getAbsoluteAbundance(object)

\S4method{getAbsoluteAbundance}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A data.frame containing absolute abundance data.
}
\description{
Accessor for AbsoluteAbundance slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getAbsoluteAbundance(CrcBiomeScreen)}: Retrieve absolute abundance data from a CrcBiomeScreen object.

}}
\examples{
getAbsoluteAbundance(CrcBiomeScreenObject)
}
