% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getModelData}
\alias{getModelData}
\title{Accessor for ModelData slot of CrcBiomeScreen object}
\usage{
getModelData(object)
}
\arguments{
\item{object}{A A \code{CrcBiomeScreen} object. object.}
}
\value{
A A \code{CrcBiomeScreen} object. object with
a data.frame containing model data.
}
\description{
Accessor for ModelData slot of CrcBiomeScreen object
}
\examples{
rel_abund <- data.frame(
  S1 = c(10),
  S2 = c(20)
)
rownames(rel_abund) <- "TaxaA"

toy_taxa <- data.frame(
  Taxa = "TaxaA",
  stringsAsFactors = FALSE
)

toy_sample <- data.frame(
  number_reads = c(10000, 10000),
  condition = c("control", "CRC"),
  row.names = c("S1", "S2"),
  stringsAsFactors = FALSE
)

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = toy_taxa,
  SampleData = toy_sample
)

getModelData(toy_obj)

}
