% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getModelData}
\alias{getModelData}
\title{Accessor for ModelData slot of CrcBiomeScreen object}
\usage{
getModelData(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a data.frame containing model data.
}
\description{
Accessor for ModelData slot of CrcBiomeScreen object
}
\examples{
rel_abund <- data.frame(Sample1 = 10, Sample2 = 20, Sample3 = 15, Sample4 = 25)
rownames(rel_abund) <- "TaxaA"

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA", row.names = "TaxaA"),
  SampleData = data.frame(
    number_reads = c(10000, 10000, 10000, 10000),
    condition = c("control", "CRC", "control", "CRC"),
    row.names = c("Sample1", "Sample2", "Sample3", "Sample4")
  )
)

slot(toy_obj, "NormalizedData") <- data.frame(
  TaxaA = c(10, 20, 15, 25),
  TaxaB = c(15, 25, 20, 30),
  row.names = c("Sample1", "Sample2", "Sample3", "Sample4")
)

toy_obj <- SplitDataSet(
  toy_obj,
  label = c("control", "CRC"),
  condition_col = "condition",
  partition = 0.7
)
getModelData(toy_obj)

}
