% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getModelData}
\alias{getModelData}
\title{Accessor for ModelData slot of CrcBiomeScreen object}
\usage{
getModelData(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a data.frame containing model data.
}
\description{
Accessor for ModelData slot of CrcBiomeScreen object
}
\examples{
toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = data.frame(TaxaA = c(10, 20)),
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = data.frame(
<<<<<<< Updated upstream
    number_reads = 10000,
    condition = "control"
  )
)
toy_obj <- SplitDataSet(
CrcBiomeScreenObject,
label = c("control", "CRC"),
partition = 0.7
=======
    number_reads = c(10000, 10000),
    condition = c("control", "CRC")
  )
)
toy_obj <- SplitDataSet(
  toy_obj,
  label = c("control", "CRC"),
  partition = 0.7
>>>>>>> Stashed changes
)
getModelData(toy_obj)

}
