% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getModelData}
\alias{getModelData}
\title{Accessor for ModelData slot of CrcBiomeScreen object}
\usage{
getModelData(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A data.frame containing model data.
}
\description{
Accessor for ModelData slot of CrcBiomeScreen object
}
\examples{
getModelData(CrcBiomeScreenObject)
}
