% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getModelResult}
\alias{getModelResult}
\alias{getModelResult,CrcBiomeScreen-method}
\title{Accessor for ModelResult slot of CrcBiomeScreen object}
\usage{
getModelResult(object)

\S4method{getModelResult}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A A \code{CrcBiomeScreen} object. object.}
}
\value{
A A \code{CrcBiomeScreen} object. object with
a list containing fitted model results.
}
\description{
Accessor for ModelResult slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getModelResult(CrcBiomeScreen)}: Retrieve model results
from a CrcBiomeScreen object.

}}
\examples{
toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = data.frame(TaxaA = c(10, 20)),
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = data.frame(
    number_reads = 10000,
    condition = "control"
  )
)

getModelData(toy_obj)
}
