% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getModelResult}
\alias{getModelResult}
\alias{getModelResult,CrcBiomeScreen-method}
\title{Accessor for ModelResult slot of CrcBiomeScreen object}
\usage{
getModelResult(object)

\S4method{getModelResult}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A list containing fitted model results.
}
\description{
Accessor for ModelResult slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getModelResult(CrcBiomeScreen)}: Retrieve model results from a CrcBiomeScreen object.

}}
\examples{
getModelResult(CrcBiomeScreenObject)
}
