% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getNormalizedData}
\alias{getNormalizedData}
\alias{getNormalizedData,CrcBiomeScreen-method}
\title{Accessor for NormalizedData slot of CrcBiomeScreen object}
\usage{
getNormalizedData(object)

\S4method{getNormalizedData}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a  data.frame (or matrix) containing normalized abundance data.
}
\description{
Retrieve normalized abundance data
from a \linkS4class{CrcBiomeScreen} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{getNormalizedData(CrcBiomeScreen)}: Retrieve normalized abundance data.

}}
\examples{
toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = data.frame(TaxaA = c(10, 20)),
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = data.frame(
    number_reads = 10000,
    condition = "control"
  )
)

getNormalizedData(toy_obj)
}
