% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getNormalizedData}
\alias{getNormalizedData}
\alias{getNormalizedData,CrcBiomeScreen-method}
\title{Accessor for NormalizedData slot of CrcBiomeScreen object}
\usage{
getNormalizedData(object)

\S4method{getNormalizedData}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A data.frame (or matrix) containing normalized abundance data.
}
\description{
Retrieve normalized abundance data from a \linkS4class{CrcBiomeScreen} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{getNormalizedData(CrcBiomeScreen)}: Retrieve normalized abundance data.

}}
\examples{
getNormalizedData(CrcBiomeScreenObject)
}
