% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getOutlierSamples}
\alias{getOutlierSamples}
\alias{getOutlierSamples,CrcBiomeScreen-method}
\title{Accessor for OutlierSamples slot of CrcBiomeScreen object}
\usage{
getOutlierSamples(object)

\S4method{getOutlierSamples}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a list containing OutlierSamples results.
}
\description{
Accessor for OutlierSamples slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getOutlierSamples(CrcBiomeScreen)}: results
from a CrcBiomeScreen object.

}}
\examples{
toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = data.frame(
    S1 = 10,
    S2 = 20,
    row.names = "TaxaA"
  ),
  TaxaData = data.frame(
    Taxa = "TaxaA",
    row.names = "TaxaA"
  ),
  SampleData = data.frame(
    number_reads = c(10000, 2000),
    condition = c("control", "control"),
    row.names = c("S1", "S2")
  )
)

slot(toy_obj, "OutlierSamples") <- c("S1")
getOutlierSamples(toy_obj)
}
