% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getPredictResult}
\alias{getPredictResult}
\alias{getPredictResult,CrcBiomeScreen-method}
\title{Accessor for PredictResult slot of CrcBiomeScreen object}
\usage{
getPredictResult(object)

\S4method{getPredictResult}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A \linkS4class{CrcBiomeScreen} object with
a  list containing fitted Prediction results.
}
\description{
Accessor for PredictResult slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getPredictResult(CrcBiomeScreen)}: Prediction results
from a CrcBiomeScreen object.

}}
\examples{
rel_abund <- data.frame(Sample1 = 10, Sample2 = 20)
rownames(rel_abund) <- "TaxaA"

toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = rel_abund,
  TaxaData = data.frame(Taxa = "TaxaA", row.names = "TaxaA"),
  SampleData = data.frame(
    number_reads = c(10000, 10000),
    condition = c("control", "CRC"),
    row.names = c("Sample1", "Sample2")
  )
)

getPredictResult(toy_obj)
}
