% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getPredictResult}
\alias{getPredictResult}
\alias{getPredictResult,CrcBiomeScreen-method}
\title{Accessor for PredictResult slot of CrcBiomeScreen object}
\usage{
getPredictResult(object)

\S4method{getPredictResult}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A list containing fitted Prediction results.
}
\description{
Accessor for PredictResult slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getPredictResult(CrcBiomeScreen)}: Prediction results from a CrcBiomeScreen object.

}}
\examples{
getPredictResult(CrcBiomeScreenObject)
}
