% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getRelativeAbundance}
\alias{getRelativeAbundance}
\alias{getRelativeAbundance,CrcBiomeScreen-method}
\title{Accessor for RelativeAbundance slot of CrcBiomeScreen object}
\usage{
getRelativeAbundance(object)

\S4method{getRelativeAbundance}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A data.frame containing relative abundance data.
}
\description{
Accessor for RelativeAbundance slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getRelativeAbundance(CrcBiomeScreen)}: Retrieve relative abundance data from a CrcBiomeScreen object.

}}
\examples{
getRelativeAbundance(CrcBiomeScreenObject)
}
