% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getSampleData}
\alias{getSampleData}
\alias{getSampleData,CrcBiomeScreen-method}
\alias{getModelData,CrcBiomeScreen-method}
\title{Accessor for SampleData slot of CrcBiomeScreen object}
\usage{
getSampleData(object)

\S4method{getSampleData}{CrcBiomeScreen}(object)

\S4method{getModelData}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A A \code{CrcBiomeScreen} object. object.}
}
\value{
A A \code{CrcBiomeScreen} object. object with
a data.frame containing sample metadata.
}
\description{
Accessor for SampleData slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getSampleData(CrcBiomeScreen)}: Retrieve sample metadata
from a CrcBiomeScreen object.

\item \code{getModelData(CrcBiomeScreen)}: Retrieve sample metadata
from a CrcBiomeScreen object.

}}
\examples{
toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = data.frame(TaxaA = c(10, 20)),
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = data.frame(
    number_reads = 10000,
    condition = "control"
  )
)

getSampleData(toy_obj)
}
