% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getSampleData}
\alias{getSampleData}
\alias{getSampleData,CrcBiomeScreen-method}
\alias{getModelData,CrcBiomeScreen-method}
\title{Accessor for SampleData slot of CrcBiomeScreen object}
\usage{
getSampleData(object)

\S4method{getSampleData}{CrcBiomeScreen}(object)

\S4method{getModelData}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A data.frame containing sample metadata.
}
\description{
Accessor for SampleData slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getSampleData(CrcBiomeScreen)}: Retrieve sample metadata from a CrcBiomeScreen object.

\item \code{getModelData(CrcBiomeScreen)}: Retrieve sample metadata from a CrcBiomeScreen object.

}}
\examples{
getSampleData(CrcBiomeScreenObject)
}
