% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{getTaxaData}
\alias{getTaxaData}
\alias{getTaxaData,CrcBiomeScreen-method}
\title{Accessor for TaxaData slot of CrcBiomeScreen object}
\usage{
getTaxaData(object)

\S4method{getTaxaData}{CrcBiomeScreen}(object)
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}
}
\value{
A data.frame containing taxonomic annotations.
}
\description{
Accessor for TaxaData slot of CrcBiomeScreen object
}
\section{Methods (by class)}{
\itemize{
\item \code{getTaxaData(CrcBiomeScreen)}: Retrieve taxonomic annotations from a CrcBiomeScreen object.

}}
\examples{
getTaxaData(CrcBiomeScreenObject)
}
