% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeepTaxonomicLevel.R
\name{KeepTaxonomicLevel}
\alias{KeepTaxonomicLevel}
\title{Keep a specific taxonomic level}
\usage{
KeepTaxonomicLevel(CrcBiomeScreenObject, level = "Genus")
}
\arguments{
\item{CrcBiomeScreenObject}{A list or object containing 'AbsoluteAbundance' and 'TaxaData'.}

\item{level}{The taxonomic level to aggregate to (e.g., "Family", "Genus", "Species").}
}
\value{
The CrcBiomeScreenObject with a new data frame aggregated at the specified level.
}
\description{
This function aggregates abundance data to a specified taxonomic level.
}
