% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadTaxaTable.R
\name{LoadTaxaTable}
\alias{LoadTaxaTable}
\title{Load a custom taxa table for ASV/OTU data}
\usage{
LoadTaxaTable(CrcBiomeScreenObject, taxa_table, id_column, taxa_column)
}
\arguments{
\item{CrcBiomeScreenObject}{The CrcBiomeScreenObject to which the taxa table will be added.}

\item{taxa_table}{A data frame. It must contain at least two columns: one for sequence IDs (e.g., ASV or OTU names)
and another for the corresponding taxonomic lineage string.}

\item{id_column}{The name of the column in \code{taxa_table} that contains the sequence IDs (ASV/OTU names).}

\item{taxa_column}{The name of the column in \code{taxa_table} that contains the taxonomic lineage string.}
}
\value{
The CrcBiomeScreenObject with the loaded taxa table.
}
\description{
This function allows users to load their own taxonomic assignments for ASV/OTU data.
The input table should map sequence IDs to their full taxonomic lineage.
}
\examples{
CrcBiomeScreenObject <- LoadTaxaTable(CrcBiomeScreenObject,
                                       taxa_table = my_taxa_table,
                                       id_column = "ASV_ID",
                                       taxa_column = "Taxonomy")
}
