% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadTaxaTable.R
\name{LoadTaxaTable}
\alias{LoadTaxaTable}
\title{Load a custom taxa table for ASV/OTU data}
\usage{
LoadTaxaTable(CrcBiomeScreenObject, taxa_table, id_column, taxa_column)
}
\arguments{
\item{CrcBiomeScreenObject}{The CrcBiomeScreenObject to which the taxa table will be added.}

\item{taxa_table}{A data frame. It must contain at least two columns: one for sequence IDs (e.g., ASV or OTU names)
and another for the corresponding taxonomic lineage string.}

\item{id_column}{The name of the column in \code{taxa_table} that contains the sequence IDs (ASV/OTU names).}

\item{taxa_column}{The name of the column in \code{taxa_table} that contains the taxonomic lineage string.}
}
\value{
The CrcBiomeScreenObject with the loaded taxa table.
}
\description{
This function allows users to load their own taxonomic assignments for ASV/OTU data.
The input table should map sequence IDs to their full taxonomic lineage.
}
\examples{
# Assuming you have a toy data object and a taxa table like this:
# CrcBiomeScreenObject <- CreateCrcBiomeScreenObject(...)
# my_taxa_table <- data.frame(
#   ASV_ID = c("ASV1", "ASV2", "ASV3"),
#   Taxonomy = c("k__Bacteria;p__Firmicutes;g__GenusA",
#                "k__Bacteria;p__Bacteroidetes;g__GenusB",
#                "k__Bacteria;p__Firmicutes;g__GenusC")
# )
# CrcBiomeScreenObject <- LoadTaxaTable(CrcBiomeScreenObject,
#                                       taxa_table = my_taxa_table,
#                                       id_column = "ASV_ID",
#                                       taxa_column = "Taxonomy")

}
