% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingRF_noweights.R
\name{ModelingRF_noweights}
\alias{ModelingRF_noweights}
\title{The function for modeling random forest without using class weights}
\usage{
ModelingRF_noweights(
  CrcBiomeScreenObject = NULL,
  k.rf = n_cv,
  TaskName = NULL,
  TrueLabel = NULL,
  num_cores = NULL
)
}
\arguments{
\item{CrcBiomeScreenObject}{A \code{CrcBiomeScreenObject} containing normalized microbiome data, sample metadata, etc.}

\item{k.rf}{Set the number of cross validation}

\item{TaskName}{A character string used to label the output}

\item{TrueLabel}{This label is the future prediction target}

\item{num_cores}{Set the number of the cores in parallel computing}
}
\value{
CrcBiomeScreenObject
}
\description{
The function for modeling random forest without using class weights
}
\examples{
CrcBiomeScreenObject <- ModelingRF_noweights(
  CrcBiomeScreenObject = CrcBiomeScreenObject,
  k.rf = n_cv,
  TaskName = TaskName,
  TrueLabel = TrueLabel,
  num_cores = num_cores
)

}
