% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingXGBoost_noweights.R
\name{ModelingXGBoost_noweights}
\alias{ModelingXGBoost_noweights}
\title{The packaging function for XGBoost modeling without using class weights}
\usage{
ModelingXGBoost_noweights(
  CrcBiomeScreenObject = NULL,
  k.rf = 10,
  repeats = 5,
  TaskName = NULL,
  TrueLabel = NULL,
  num_cores = num_cores
)
}
\arguments{
\item{CrcBiomeScreenObject}{A \code{CrcBiomeScreenObject} containing normalized microbiome data, sample metadata, etc.}

\item{k.rf}{Set the number of cross validation}

\item{repeats}{Set the number of repeats in cross validation}

\item{TaskName}{A character string used to label the output}

\item{TrueLabel}{This label is the future prediction target}

\item{num_cores}{Set the number of the cores in parallel computing}
}
\value{
CrcBiomeScreenObject
}
\description{
The packaging function for XGBoost modeling without using class weights
}
\examples{
CrcBiomeScreenObject <- ModelingXGBoost_noweights(
                                  CrcBiomeScreenObject = CrcBiomeScreenObject,
                                  k.rf = n_cv,
                                  TaskName = TaskName,
                                  TrueLabel = TrueLabel,
                                  num_cores = num_cores)
}
