% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSBCSP_screeningData.R
\docType{data}
\name{NHSBCSP_screeningData}
\alias{NHSBCSP_screeningData}
\title{NHSBCSP screening dataset (toy example)}
\format{
A data.frame
}
\source{
NHS Bowel Cancer Screening Programme
}
\usage{
NHSBCSP_screeningData
}
\description{
The NHS Bowel Cancer Screening Programme microbiome data.
A dataset included in the package.
}
\keyword{datasets}
