% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeData.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalise the absolute data to relative data by using Total sum scaling
and Geometric Mean of Pairwise Ratios (GMPR)}
\usage{
NormalizeData(CrcBiomeScreenObject = NULL, method = NULL)
}
\arguments{
\item{CrcBiomeScreenObject}{From the CreateCrcBiomeScreenObject()}

\item{method}{TSS or GMPR}
}
\value{
CrcBiomeScreenObject$NormalizedData
}
\description{
Normalise the absolute data to relative data by using Total sum scaling
and Geometric Mean of Pairwise Ratios (GMPR)
}
\examples{
CrcBiomeScreenObject <- NormalizeData(CrcBiomeScreenObject,
                                                method = "GMPR",
                                                TaskName = "Normalize_ToyData")
          attributes(CrcBiomeScreenObject$NormalizedData)

}
