% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeData.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalise the absolute data to relative data by using Total Sum Scaling
and Geometric Mean of Pairwise Ratios (GMPR)}
\usage{
NormalizeData(CrcBiomeScreenObject = NULL, method = NULL, level = NULL)
}
\arguments{
\item{CrcBiomeScreenObject}{From the CreateCrcBiomeScreenObject()}

\item{method}{"TSS" or "GMPR"}
}
\value{
Updated CrcBiomeScreenObject with NormalizedData
}
\description{
Normalise the absolute data to relative data by using Total Sum Scaling
and Geometric Mean of Pairwise Ratios (GMPR)
}
\examples{
# Normalize using GMPR (Geometric Mean of Pairwise Ratios)
CrcBiomeScreenObject <- NormalizeData(CrcBiomeScreenObject, method = "GMPR", level = "Genus")
# Normalize using TSS (Total Sum Scaling)
CrcBiomeScreenObject <- NormalizeData(CrcBiomeScreenObject, method = "TSS", level = "Genus")
}
