% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCrcBiomeScreen.R
\name{PredictCrcBiomeScreen}
\alias{PredictCrcBiomeScreen}
\title{Predict the class and probabilities for new data}
\usage{
PredictCrcBiomeScreen(
  CrcBiomeScreenObject,
  newdata,
  model_type = c("RF", "XGBoost")
)
}
\arguments{
\item{CrcBiomeScreenObject}{The object containing the trained model.}

\item{newdata}{The data frame or matrix of new features to predict on.}

\item{model_type}{The type of model to use for prediction ("RF" or "XGBoost").}
}
\value{
A data frame containing sample-specific predictions.
}
\description{
Predict the class and probabilities for new data
}
