% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunScreening.R
\name{RunScreening}
\alias{RunScreening}
\title{Run the screening process for the microbiome data}
\usage{
RunScreening(
  obj,
  model_type = NULL,
  split.requirement = NULL,
  TaskName = TaskName,
  partition = NULL,
  ClassWeights = NULL,
  n_cv = NULL,
  ValidationData = NULL,
  TrueLabel = NULL,
  num_cores = NULL
)
}
\arguments{
\item{obj}{A \code{CrcBiomeScreenObject} containing normalized microbiome data, sample metadata, etc.}

\item{model_type}{Model type to be used, default is "RF"}

\item{split.requirement}{A list containing the label and condition column for splitting the dataset, default is NULL}

\item{TaskName}{A character string used to label the output}

\item{partition}{The number of partitions for cross-validation}

\item{ClassWeights}{Whether to use class weights in the model training, default is NULL}

\item{n_cv}{The number of cross-validation folds, default is NULL}

\item{ValidationData}{A \code{CrcBiomeScreenObject} containing validation data for model evaluation, default is NULL}

\item{TrueLabel}{The true label for the classification task, which is used to evaluate the model's performance}

\item{num_cores}{Set the number of cores for parallel computing, default is NULL}

\item{normalize_method}{Normalization method to be used, default is "TSS"}
}
\value{
A \code{CrcBiomeScreenObject} with the results of the screening process, including model training, evaluation, and validation.
}
\description{
Run the screening process for the microbiome data
}
\examples{
CrcBiomeScreenObject <- RunScreening(CrcBiomeScreenObject,
  normalize_method = "GMPR",
  model = "RF",
  split.requirement =
    c(
      label = c("control", "CRC"),
      condition_col = "study_condition"
    ),
  TaskName = "GMPR_NHSBCSP",
  num_cores = 10,
  ValidationData = ValidationData,
  TrueLabel = "Cancer"
)
}
