% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{setNormalizedData-setter}
\alias{setNormalizedData-setter}
\alias{setNormalizedData<-}
\alias{setNormalizedData<-,CrcBiomeScreen-method}
\title{setNormalizedData<-: Setter for NormalizedData slot of CrcBiomeScreen object}
\usage{
setNormalizedData(object) <- value

\S4method{setNormalizedData}{CrcBiomeScreen}(object) <- value
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}

\item{value}{A data.frame or matrix containing normalized abundance data.}
}
\value{
The modified \linkS4class{CrcBiomeScreen} object.
}
\description{
setNormalizedData<-: Setter for NormalizedData slot of CrcBiomeScreen object
}
\section{Functions}{
\itemize{
\item \code{setNormalizedData(CrcBiomeScreen) <- value}: Replace the NormalizedData slot of a CrcBiomeScreen object.

}}
\examples{
setNormalizedData(CrcBiomeScreenObject)
}
