% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{setTaxaData-setter}
\alias{setTaxaData-setter}
\alias{setTaxaData<-}
\alias{setTaxaData<-,CrcBiomeScreen-method}
\title{setTaxaData<-: Setter for TaxaData slot of CrcBiomeScreen object}
\usage{
setTaxaData(object) <- value

\S4method{setTaxaData}{CrcBiomeScreen}(object) <- value
}
\arguments{
\item{object}{A A \code{CrcBiomeScreen} object. object.}

\item{value}{A data.frame containing updated taxonomic annotations.}
}
\value{
A A \code{CrcBiomeScreen} object. object with
a modified A \code{CrcBiomeScreen} object. object.
}
\description{
setTaxaData<-: Setter for TaxaData slot of CrcBiomeScreen object
}
\section{Functions}{
\itemize{
\item \code{setTaxaData(CrcBiomeScreen) <- value}: Replace the TaxaData slot
of a CrcBiomeScreen object.

}}
\examples{
toy_obj <- CreateCrcBiomeScreenObject(
  RelativeAbundance = data.frame(TaxaA = c(10, 20)),
  TaxaData = data.frame(Taxa = "TaxaA"),
  SampleData = data.frame(
    number_reads = 10000,
    condition = "control"
  )
)

setTaxaData(toy_obj) <- data.frame(Taxa = "NewTaxa")
getTaxaData(toy_obj)
}
