% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCrcBiomeScreenObject.R
\name{setTaxaData-setter}
\alias{setTaxaData-setter}
\alias{setTaxaData<-}
\alias{setTaxaData<-,CrcBiomeScreen-method}
\title{setTaxaData<-: Setter for TaxaData slot of CrcBiomeScreen object}
\usage{
setTaxaData(object) <- value

\S4method{setTaxaData}{CrcBiomeScreen}(object) <- value
}
\arguments{
\item{object}{A \linkS4class{CrcBiomeScreen} object.}

\item{value}{A data.frame containing updated taxonomic annotations.}
}
\value{
The modified \linkS4class{CrcBiomeScreen} object.
}
\description{
setTaxaData<-: Setter for TaxaData slot of CrcBiomeScreen object
}
\section{Functions}{
\itemize{
\item \code{setTaxaData(CrcBiomeScreen) <- value}: Replace the TaxaData slot of a CrcBiomeScreen object.

}}
\examples{
setTaxaData(CrcBiomeScreenObject)
}
