% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitDataSet.R
\name{SplitDataSet}
\alias{SplitDataSet}
\title{Split the dataset into training and test sets}
\usage{
SplitDataSet(
  CrcBiomeScreenObject = NULL,
  label = NULL,
  partition = NULL,
  condition_col = "study_condition"
)
}
\arguments{
\item{CrcBiomeScreenObject}{From the CreateCrcBiomeScreenObject()}

\item{label}{Divide the data set by the binary-label}

\item{partition}{The ratio of dividing the data set}

\item{condition_col}{The colname of label in SampleData}
}
\value{
CrcBiomeScreenObject$ModelData
}
\description{
Split the dataset into training and test sets
}
\examples{
CrcBiomeScreenObject <- SplitDataSet(CrcBiomeScreenObject,
  label = c("control", "CRC"),
  partition = 0.7,
  condition_col = "study_condition"
)

}
