% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitTaxas.R
\name{SplitTaxas}
\alias{SplitTaxas}
\title{Split the taxa information into different columns}
\usage{
SplitTaxas(CrcBiomeScreenObject)
}
\arguments{
\item{CrcBiomeScreenObject}{}
}
\value{
CrcBiomeScreenObject$TaxaData
}
\description{
Split the taxa information into different columns
}
\examples{
CrcBiomeScreenObject <- SplitTaxas(CrcBiomeScreenObject)
}
