% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitTaxas.R
\name{SplitTaxas}
\alias{SplitTaxas}
\title{Split and clean taxonomy strings}
\usage{
SplitTaxas(CrcBiomeScreenObject)
}
\arguments{
\item{CrcBiomeScreenObject}{}
}
\value{
CrcBiomeScreenObject$TaxaData
}
\description{
This function automatically detects the taxonomy string format (e.g., MetaPhlAn, QIIME, SILVA, GTDB),
splits the string into standard taxonomic ranks (Kingdom to Species),
retains the original taxonomy string in a new column (\code{OriginalTaxa}),
and refines labels such as "uncultured" or "unclassified" by appending the parent rank.
}
\examples{
CrcBiomeScreenObject <- SplitTaxas(CrcBiomeScreenObject)
}
