% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_2018_RelativeAbundance.R
\docType{data}
\name{Thomas_2018_RelativeAbundance}
\alias{Thomas_2018_RelativeAbundance}
\title{Thomas_2018_RelativeAbundance}
\format{
A data frame with ...
}
\source{
curatedMetagenomicData package, Thomas et al. 2018.
A data frame (or matrix) of relative abundance values derived from
Thomas et al. (2018), where rows correspond to microbial taxa and
columns correspond to samples.
}
\usage{
Thomas_2018_RelativeAbundance
}
\description{
A dataset: Thomas_2018_RelativeAbundance microbiome data.
}
\keyword{datasets}
