% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thomas_2018_RelativeAbundance.R
\docType{data}
\name{Thomas_2018_RelativeAbundance}
\alias{Thomas_2018_RelativeAbundance}
\title{Thomas 2018 relative abundance dataset}
\format{
A named list containing relative abundance objects for demonstration.
}
\source{
curatedMetagenomicData
}
\usage{
data(Thomas_2018_RelativeAbundance)
}
\description{
A relative abundance dataset derived from curatedMetagenomicData and included
for package examples.
2021-03-31.ThomasAM_2018a.relative_abundance:
Formal class 'TreeSummarizedExperiment' from
package "TreeSummarizedExperiment" with 14 slots.
}
\keyword{datasets}
