% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModels.R
\name{TrainModels}
\alias{TrainModels}
\title{Train the different models}
\usage{
TrainModels(
  CrcBiomeScreenObject = NULL,
  model_type = c("RF", "XGBoost"),
  ClassWeights = TRUE,
  n_cv = 10,
  TaskName = NULL,
  TrueLabel = NULL,
  num_cores = NULL
)
}
\arguments{
\item{CrcBiomeScreenObject}{A \code{CrcBiomeScreenObject} containing normalized microbiome data, sample metadata, etc.}

\item{model_type}{Select the method for modeling}

\item{ClassWeights}{Choose using the class weights or not}

\item{n_cv}{Set the number of cross validation}

\item{TaskName}{A character string used to label the output}

\item{TrueLabel}{This label is the future prediction target}

\item{num_cores}{Set the number of the cores in parallel computing}
}
\value{
A \linkS4class{CrcBiomeScreenObject} with training results.
}
\description{
Train the different models
}
\examples{

train_df <- data.frame(
  FeatureA = c(1, 2, 3, 4),
  TrainLabel = factor(c("control", "CRC", "control", "CRC"))
)

toy_obj <- new("CrcBiomeScreen",
  AbsoluteAbundance = data.frame(),
  RelativeAbundance = data.frame(),
  TaxaData = data.frame(),
  SampleData = data.frame(),
  ModelData = list(
    Training = train_df,
    TrainLabel = train_df$TrainLabel
  )
)

if (interactive()) {
  toy_obj <- TrainModels(
    CrcBiomeScreenObject = toy_obj,
    model_type = "RF",
    TaskName = "ToyData_RF",
    ClassWeights = TRUE,
    TrueLabel = "CRC",
    n_cv = 2,
    num_cores = 1
  )
}

}
