% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModels.R
\name{TrainModels}
\alias{TrainModels}
\title{Train the different models}
\usage{
TrainModels(
  CrcBiomeScreenObject = NULL,
  model_type = c("RF", "XGBoost"),
  ClassBalance = TRUE,
  n_cv = 10,
  TaskName = NULL,
  TrueLabel = NULL,
  num_cores = NULL
)
}
\arguments{
\item{CrcBiomeScreenObject}{A \code{CrcBiomeScreenObject} containing normalized microbiome data, sample metadata, etc.}

\item{model_type}{Select the method for modeling}

\item{ClassBalance}{Choose using the class weights or not}

\item{n_cv}{Set the number of cross validation}

\item{TaskName}{A character string used to label the output}

\item{TrueLabel}{This label is the future prediction target}

\item{num_cores}{Set the number of the cores in parallel computing}
}
\value{
CrcBiomeScreenObject
}
\description{
Train the different models
}
\examples{
CrcBiomeScreenObject <- TrainModels(CrcBiomeScreenObject,
                                              model_type = "RF",
                                              TaskName = "ToyData_RF",
                                              ClassBalance = TRUE,
                                              TrueLabel = "CRC",
                                              num_cores = 10)

}
