% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModelOnData.R
\name{ValidateModelOnData}
\alias{ValidateModelOnData}
\title{Evaluate the model on the validation data}
\usage{
ValidateModelOnData(
  CrcBiomeScreenObject = NULL,
  model_type = NULL,
  ValidationData = NULL,
  TaskName = NULL,
  TrueLabel = NULL,
  condition_col = "study_condition",
  PlotAUC = NULL
)
}
\arguments{
\item{CrcBiomeScreenObject}{A CrcBiomeScreenObject containing the model and data to be evaluated.}

\item{model_type}{The type of model to be evaluated, either "RF" for Random Forest or "XGBoost".}

\item{ValidationData}{A CrcBiomeScreenObject containing the validation data to be used for model evaluation.}

\item{TaskName}{A character string used to label the output files and results.}

\item{TrueLabel}{The true label for the classification task, which is used to evaluate the model's performance.}

\item{condition_col}{The column name in the SampleData that contains the study condition labels. Default is "study_condition".}

\item{PlotAUC}{A logical value indicating whether to plot the AUC curve. If TRUE, the AUC curve will be saved as a PDF file.}
}
\value{
A CrcBiomeScreenObject with the evaluation results stored in the \code{PredictResult} slot for the specified model type.
}
\description{
Evaluate the model on the validation data
}
\examples{
CrcBiomeScreenObject <- ValidateModelOnData(CrcBiomeScreenObject,
                                                      model_type = "RF",
                                                      ValidationData = ValidationData_filtered_qc,
                                                      TaskName = "ValidationData_RF_Validation",
                                                      TrueLabel = "CRC",
                                                      PlotAUC = TRUE)

}
