% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZellerG_2014_RelativeAbundance.R
\docType{data}
\name{ZellerG_2014_RelativeAbundance}
\alias{ZellerG_2014_RelativeAbundance}
\title{Zeller 2014 relative abundance dataset}
\format{
A named list containing relative abundance objects for demonstration.
}
\source{
curatedMetagenomicData
}
\usage{
ZellerG_2014_RelativeAbundance
}
\description{
A relative abundance dataset derived from curatedMetagenomicData and included
for package examples.
2021-03-31.ZellerG_2014.relative_abundance:
Formal class 'TreeSummarizedExperiment' from
package "TreeSummarizedExperiment" with 14 slots.
}
\keyword{datasets}
